requirements.txt
aiogram==3.17.0
Scrapy==2.12.0
python-dotenv==1.0.1
gigachat==0.1.37.post1
scikit-learn==1.6.1
pandas==2.2.3
nltk==3.9.1
aiohttp==3.11.11